refacto config fix indentation in make file add test for make build
