+++ .github/workflows/build.yml
@@ -19,6 +19,22 @@ permissions:
 
 jobs:
 
+  makefile:
+    runs-on: ubuntu-latest
+    
+    strategy:
+      matrix:
+        go-version: ['1.20', '1.21']
+
+    steps:
+      - uses: actions/checkout@v4
+      - uses: actions/setup-go@v4
+        with:
+          go-version: ${{ matrix.go-version }}
+  
+      - name: build binary from make file
+        run: make build
+      
   go-ubuntu:
     runs-on: ubuntu-latest
