+++ pkgconfig/config.go
@@ -4,7 +4,6 @@ import (
 	"os"
 
 	"github.com/dmachard/go-dnscollector/netlib"
-	"github.com/prometheus/prometheus/model/relabel"
 	"gopkg.in/yaml.v3"
 )
 
@@ -19,504 +18,13 @@ func IsValidMode(mode string) bool {
 	return false
 }
 
-type MultiplexInOut struct {
-	Name       string                 `yaml:"name"`
-	Transforms map[string]interface{} `yaml:"transforms"`
-	Params     map[string]interface{} `yaml:",inline"`
-}
-
-type MultiplexRoutes struct {
-	Src []string `yaml:"from,flow"`
-	Dst []string `yaml:"to,flow"`
-}
-
-type ConfigTransformers struct {
-	UserPrivacy struct {
-		Enable        bool `yaml:"enable"`
-		AnonymizeIP   bool `yaml:"anonymize-ip"`
-		MinimazeQname bool `yaml:"minimaze-qname"`
-		HashIP        bool `yaml:"hash-ip"`
-	} `yaml:"user-privacy"`
-	Normalize struct {
-		Enable         bool `yaml:"enable"`
-		QnameLowerCase bool `yaml:"qname-lowercase"`
-		QuietText      bool `yaml:"quiet-text"`
-		AddTld         bool `yaml:"add-tld"`
-		AddTldPlusOne  bool `yaml:"add-tld-plus-one"`
-	} `yaml:"normalize"`
-	Latency struct {
-		Enable            bool `yaml:"enable"`
-		MeasureLatency    bool `yaml:"measure-latency"`
-		UnansweredQueries bool `yaml:"unanswered-queries"`
-		QueriesTimeout    int  `yaml:"queries-timeout"`
-	}
-	Reducer struct {
-		Enable                    bool `yaml:"enable"`
-		RepetitiveTrafficDetector bool `yaml:"repetitive-traffic-detector"`
-		QnamePlusOne              bool `yaml:"qname-plus-one"`
-		WatchInterval             int  `yaml:"watch-interval"`
-	}
-	Filtering struct {
-		Enable          bool     `yaml:"enable"`
-		DropFqdnFile    string   `yaml:"drop-fqdn-file"`
-		DropDomainFile  string   `yaml:"drop-domain-file"`
-		KeepFqdnFile    string   `yaml:"keep-fqdn-file"`
-		KeepDomainFile  string   `yaml:"keep-domain-file"`
-		DropQueryIPFile string   `yaml:"drop-queryip-file"`
-		KeepQueryIPFile string   `yaml:"keep-queryip-file"`
-		KeepRdataFile   string   `yaml:"keep-rdata-file"`
-		DropRcodes      []string `yaml:"drop-rcodes,flow"`
-		LogQueries      bool     `yaml:"log-queries"`
-		LogReplies      bool     `yaml:"log-replies"`
-		Downsample      int      `yaml:"downsample"`
-	} `yaml:"filtering"`
-	GeoIP struct {
-		Enable        bool   `yaml:"enable"`
-		DBCountryFile string `yaml:"mmdb-country-file"`
-		DBCityFile    string `yaml:"mmdb-city-file"`
-		DBASNFile     string `yaml:"mmdb-asn-file"`
-	} `yaml:"geoip"`
-	Suspicious struct {
-		Enable             bool     `yaml:"enable"`
-		ThresholdQnameLen  int      `yaml:"threshold-qname-len"`
-		ThresholdPacketLen int      `yaml:"threshold-packet-len"`
-		ThresholdSlow      float64  `yaml:"threshold-slow"`
-		CommonQtypes       []string `yaml:"common-qtypes,flow"`
-		UnallowedChars     []string `yaml:"unallowed-chars,flow"`
-		ThresholdMaxLabels int      `yaml:"threshold-max-labels"`
-		WhitelistDomains   []string `yaml:"whitelist-domains,flow"`
-	} `yaml:"suspicious"`
-	Extract struct {
-		Enable     bool `yaml:"enable"`
-		AddPayload bool `yaml:"add-payload"`
-	} `yaml:"extract"`
-	MachineLearning struct {
-		Enable      bool `yaml:"enable"`
-		AddFeatures bool `yaml:"add-features"`
-	} `yaml:"machine-learning"`
-}
-
-func (c *ConfigTransformers) SetDefault() {
-	c.Suspicious.Enable = false
-	c.Suspicious.ThresholdQnameLen = 100
-	c.Suspicious.ThresholdPacketLen = 1000
-	c.Suspicious.ThresholdSlow = 1.0
-	c.Suspicious.CommonQtypes = []string{"A", "AAAA", "TXT", "CNAME", "PTR",
-		"NAPTR", "DNSKEY", "SRV", "SOA", "NS", "MX", "DS", "HTTPS"}
-	c.Suspicious.UnallowedChars = []string{"\"", "==", "/", ":"}
-	c.Suspicious.ThresholdMaxLabels = 10
-	c.Suspicious.WhitelistDomains = []string{"\\.ip6\\.arpa"}
-
-	c.UserPrivacy.Enable = false
-	c.UserPrivacy.AnonymizeIP = false
-	c.UserPrivacy.MinimazeQname = false
-	c.UserPrivacy.HashIP = false
-
-	c.Normalize.Enable = false
-	c.Normalize.QnameLowerCase = false
-	c.Normalize.QuietText = false
-	c.Normalize.AddTld = false
-	c.Normalize.AddTldPlusOne = false
-
-	c.Latency.Enable = false
-	c.Latency.MeasureLatency = false
-	c.Latency.UnansweredQueries = false
-	c.Latency.QueriesTimeout = 2
-
-	c.Reducer.Enable = false
-	c.Reducer.RepetitiveTrafficDetector = false
-	c.Reducer.QnamePlusOne = false
-	c.Reducer.WatchInterval = 5
-
-	c.Filtering.Enable = false
-	c.Filtering.DropFqdnFile = ""
-	c.Filtering.DropDomainFile = ""
-	c.Filtering.KeepFqdnFile = ""
-	c.Filtering.KeepDomainFile = ""
-	c.Filtering.DropQueryIPFile = ""
-	c.Filtering.DropRcodes = []string{}
-	c.Filtering.LogQueries = true
-	c.Filtering.LogReplies = true
-	c.Filtering.Downsample = 0
-
-	c.GeoIP.Enable = false
-	c.GeoIP.DBCountryFile = ""
-	c.GeoIP.DBCityFile = ""
-	c.GeoIP.DBASNFile = ""
-
-	c.Extract.Enable = false
-	c.Extract.AddPayload = false
-
-	c.MachineLearning.Enable = false
-	c.MachineLearning.AddFeatures = false
-}
-
-/* main configuration */
 type Config struct {
-	Global struct {
-		TextFormat          string `yaml:"text-format"`
-		TextFormatDelimiter string `yaml:"text-format-delimiter"`
-		TextFormatBoundary  string `yaml:"text-format-boundary"`
-		Trace               struct {
-			Verbose      bool   `yaml:"verbose"`
-			LogMalformed bool   `yaml:"log-malformed"`
-			Filename     string `yaml:"filename"`
-			MaxSize      int    `yaml:"max-size"`
-			MaxBackups   int    `yaml:"max-backups"`
-		} `yaml:"trace"`
-		ServerIdentity string `yaml:"server-identity"`
-	} `yaml:"global"`
-
-	Collectors struct {
-		Tail struct {
-			Enable       bool   `yaml:"enable"`
-			TimeLayout   string `yaml:"time-layout"`
-			PatternQuery string `yaml:"pattern-query"`
-			PatternReply string `yaml:"pattern-reply"`
-			FilePath     string `yaml:"file-path"`
-		} `yaml:"tail"`
-		Dnstap struct {
-			Enable            bool   `yaml:"enable"`
-			ListenIP          string `yaml:"listen-ip"`
-			ListenPort        int    `yaml:"listen-port"`
-			SockPath          string `yaml:"sock-path"`
-			TLSSupport        bool   `yaml:"tls-support"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			RcvBufSize        int    `yaml:"sock-rcvbuf"`
-			ResetConn         bool   `yaml:"reset-conn"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-			DisableDNSParser  bool   `yaml:"disable-dnsparser"`
-		} `yaml:"dnstap"`
-		DnstapProxifier struct {
-			Enable        bool   `yaml:"enable"`
-			ListenIP      string `yaml:"listen-ip"`
-			ListenPort    int    `yaml:"listen-port"`
-			SockPath      string `yaml:"sock-path"`
-			TLSSupport    bool   `yaml:"tls-support"`
-			TLSMinVersion string `yaml:"tls-min-version"`
-			CertFile      string `yaml:"cert-file"`
-			KeyFile       string `yaml:"key-file"`
-		} `yaml:"dnstap-proxifier"`
-		AfpacketLiveCapture struct {
-			Enable            bool   `yaml:"enable"`
-			Port              int    `yaml:"port"`
-			Device            string `yaml:"device"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"afpacket-sniffer"`
-		XdpLiveCapture struct {
-			Enable            bool   `yaml:"enable"`
-			Port              int    `yaml:"port"`
-			Device            string `yaml:"device"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"xdp-sniffer"`
-		PowerDNS struct {
-			Enable            bool   `yaml:"enable"`
-			ListenIP          string `yaml:"listen-ip"`
-			ListenPort        int    `yaml:"listen-port"`
-			TLSSupport        bool   `yaml:"tls-support"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			AddDNSPayload     bool   `yaml:"add-dns-payload"`
-			RcvBufSize        int    `yaml:"sock-rcvbuf"`
-			ResetConn         bool   `yaml:"reset-conn"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"powerdns"`
-		FileIngestor struct {
-			Enable            bool   `yaml:"enable"`
-			WatchDir          string `yaml:"watch-dir"`
-			WatchMode         string `yaml:"watch-mode"`
-			PcapDNSPort       int    `yaml:"pcap-dns-port"`
-			DeleteAfter       bool   `yaml:"delete-after"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"file-ingestor"`
-		Tzsp struct {
-			Enable            bool   `yaml:"enable"`
-			ListenIP          string `yaml:"listen-ip"`
-			ListenPort        int    `yaml:"listen-port"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"tzsp"`
-	} `yaml:"collectors"`
-
-	IngoingTransformers ConfigTransformers `yaml:"collectors-transformers"`
-
-	Loggers struct {
-		Stdout struct {
-			Enable            bool   `yaml:"enable"`
-			Mode              string `yaml:"mode"`
-			TextFormat        string `yaml:"text-format"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"stdout"`
-		Prometheus struct {
-			Enable                  bool     `yaml:"enable"`
-			ListenIP                string   `yaml:"listen-ip"`
-			ListenPort              int      `yaml:"listen-port"`
-			TLSSupport              bool     `yaml:"tls-support"`
-			TLSMutual               bool     `yaml:"tls-mutual"`
-			TLSMinVersion           string   `yaml:"tls-min-version"`
-			CertFile                string   `yaml:"cert-file"`
-			KeyFile                 string   `yaml:"key-file"`
-			PromPrefix              string   `yaml:"prometheus-prefix"`
-			LabelsList              []string `yaml:"prometheus-labels"`
-			TopN                    int      `yaml:"top-n"`
-			BasicAuthLogin          string   `yaml:"basic-auth-login"`
-			BasicAuthPwd            string   `yaml:"basic-auth-pwd"`
-			BasicAuthEnabled        bool     `yaml:"basic-auth-enable"`
-			ChannelBufferSize       int      `yaml:"chan-buffer-size"`
-			HistogramMetricsEnabled bool     `yaml:"histogram-metrics-enabled"`
-		} `yaml:"prometheus"`
-		RestAPI struct {
-			Enable            bool   `yaml:"enable"`
-			ListenIP          string `yaml:"listen-ip"`
-			ListenPort        int    `yaml:"listen-port"`
-			BasicAuthLogin    string `yaml:"basic-auth-login"`
-			BasicAuthPwd      string `yaml:"basic-auth-pwd"`
-			TLSSupport        bool   `yaml:"tls-support"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			TopN              int    `yaml:"top-n"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"restapi"`
-		LogFile struct {
-			Enable              bool   `yaml:"enable"`
-			FilePath            string `yaml:"file-path"`
-			MaxSize             int    `yaml:"max-size"`
-			MaxFiles            int    `yaml:"max-files"`
-			FlushInterval       int    `yaml:"flush-interval"`
-			Compress            bool   `yaml:"compress"`
-			CompressInterval    int    `yaml:"compress-interval"`
-			CompressPostCommand string `yaml:"compress-postcommand"`
-			Mode                string `yaml:"mode"`
-			PostRotateCommand   string `yaml:"postrotate-command"`
-			PostRotateDelete    bool   `yaml:"postrotate-delete-success"`
-			TextFormat          string `yaml:"text-format"`
-			ChannelBufferSize   int    `yaml:"chan-buffer-size"`
-		} `yaml:"logfile"`
-		DNSTap struct {
-			Enable            bool   `yaml:"enable"`
-			RemoteAddress     string `yaml:"remote-address"`
-			RemotePort        int    `yaml:"remote-port"`
-			Transport         string `yaml:"transport"`
-			SockPath          string `yaml:"sock-path"`
-			ConnectTimeout    int    `yaml:"connect-timeout"`
-			RetryInterval     int    `yaml:"retry-interval"`
-			FlushInterval     int    `yaml:"flush-interval"`
-			TLSSupport        bool   `yaml:"tls-support"`
-			TLSInsecure       bool   `yaml:"tls-insecure"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CAFile            string `yaml:"ca-file"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			ServerID          string `yaml:"server-id"`
-			OverwriteIdentity bool   `yaml:"overwrite-identity"`
-			BufferSize        int    `yaml:"buffer-size"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"dnstap"`
-		TCPClient struct {
-			Enable            bool   `yaml:"enable"`
-			RemoteAddress     string `yaml:"remote-address"`
-			RemotePort        int    `yaml:"remote-port"`
-			SockPath          string `yaml:"sock-path"` // deprecated
-			RetryInterval     int    `yaml:"retry-interval"`
-			Transport         string `yaml:"transport"`
-			TLSSupport        bool   `yaml:"tls-support"` // deprecated
-			TLSInsecure       bool   `yaml:"tls-insecure"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CAFile            string `yaml:"ca-file"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			Mode              string `yaml:"mode"`
-			TextFormat        string `yaml:"text-format"`
-			PayloadDelimiter  string `yaml:"delimiter"`
-			BufferSize        int    `yaml:"buffer-size"`
-			FlushInterval     int    `yaml:"flush-interval"`
-			ConnectTimeout    int    `yaml:"connect-timeout"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"tcpclient"`
-		Syslog struct {
-			Enable            bool   `yaml:"enable"`
-			Severity          string `yaml:"severity"`
-			Facility          string `yaml:"facility"`
-			Transport         string `yaml:"transport"`
-			RemoteAddress     string `yaml:"remote-address"`
-			RetryInterval     int    `yaml:"retry-interval"`
-			TextFormat        string `yaml:"text-format"`
-			Mode              string `yaml:"mode"`
-			TLSInsecure       bool   `yaml:"tls-insecure"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CAFile            string `yaml:"ca-file"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			Formatter         string `yaml:"formatter"`
-			Framer            string `yaml:"framer"`
-			Hostname          string `yaml:"hostname"`
-			AppName           string `yaml:"app-name"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-			Tag               string `yaml:"tag"`
-			ReplaceNullChar   string `yaml:"replace-null-char"`
-			FlushInterval     int    `yaml:"flush-interval"`
-			BufferSize        int    `yaml:"buffer-size"`
-		} `yaml:"syslog"`
-		Fluentd struct {
-			Enable            bool   `yaml:"enable"`
-			RemoteAddress     string `yaml:"remote-address"`
-			RemotePort        int    `yaml:"remote-port"`
-			SockPath          string `yaml:"sock-path"` // deprecated
-			ConnectTimeout    int    `yaml:"connect-timeout"`
-			RetryInterval     int    `yaml:"retry-interval"`
-			FlushInterval     int    `yaml:"flush-interval"`
-			Transport         string `yaml:"transport"`
-			TLSSupport        bool   `yaml:"tls-support"` // deprecated
-			TLSInsecure       bool   `yaml:"tls-insecure"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CAFile            string `yaml:"ca-file"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			Tag               string `yaml:"tag"`
-			BufferSize        int    `yaml:"buffer-size"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"fluentd"`
-		InfluxDB struct {
-			Enable            bool   `yaml:"enable"`
-			ServerURL         string `yaml:"server-url"`
-			AuthToken         string `yaml:"auth-token"`
-			TLSSupport        bool   `yaml:"tls-support"`
-			TLSInsecure       bool   `yaml:"tls-insecure"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CAFile            string `yaml:"ca-file"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			Bucket            string `yaml:"bucket"`
-			Organization      string `yaml:"organization"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"influxdb"`
-		LokiClient struct {
-			Enable            bool              `yaml:"enable"`
-			ServerURL         string            `yaml:"server-url"`
-			JobName           string            `yaml:"job-name"`
-			Mode              string            `yaml:"mode"`
-			FlushInterval     int               `yaml:"flush-interval"`
-			BatchSize         int               `yaml:"batch-size"`
-			RetryInterval     int               `yaml:"retry-interval"`
-			TextFormat        string            `yaml:"text-format"`
-			ProxyURL          string            `yaml:"proxy-url"`
-			TLSInsecure       bool              `yaml:"tls-insecure"`
-			TLSMinVersion     string            `yaml:"tls-min-version"`
-			CAFile            string            `yaml:"ca-file"`
-			CertFile          string            `yaml:"cert-file"`
-			KeyFile           string            `yaml:"key-file"`
-			BasicAuthLogin    string            `yaml:"basic-auth-login"`
-			BasicAuthPwd      string            `yaml:"basic-auth-pwd"`
-			BasicAuthPwdFile  string            `yaml:"basic-auth-pwd-file"`
-			TenantID          string            `yaml:"tenant-id"`
-			RelabelConfigs    []*relabel.Config `yaml:"relabel-configs"`
-			ChannelBufferSize int               `yaml:"chan-buffer-size"`
-		} `yaml:"lokiclient"`
-		Statsd struct {
-			Enable            bool   `yaml:"enable"`
-			Prefix            string `yaml:"prefix"`
-			RemoteAddress     string `yaml:"remote-address"`
-			RemotePort        int    `yaml:"remote-port"`
-			ConnectTimeout    int    `yaml:"connect-timeout"`
-			Transport         string `yaml:"transport"`
-			FlushInterval     int    `yaml:"flush-interval"`
-			TLSSupport        bool   `yaml:"tls-support"` // deprecated
-			TLSInsecure       bool   `yaml:"tls-insecure"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CAFile            string `yaml:"ca-file"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"statsd"`
-		ElasticSearchClient struct {
-			Enable            bool   `yaml:"enable"`
-			Index             string `yaml:"index"`
-			Server            string `yaml:"server"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-			BulkSize          int    `yaml:"bulk-size"`
-			FlushInterval     int    `yaml:"flush-interval"`
-		} `yaml:"elasticsearch"`
-		ScalyrClient struct {
-			Enable            bool                   `yaml:"enable"`
-			Mode              string                 `yaml:"mode"`
-			TextFormat        string                 `yaml:"text-format"`
-			SessionInfo       map[string]string      `yaml:"sessioninfo"`
-			Attrs             map[string]interface{} `yaml:"attrs"`
-			ServerURL         string                 `yaml:"server-url"`
-			APIKey            string                 `yaml:"apikey"`
-			Parser            string                 `yaml:"parser"`
-			FlushInterval     int                    `yaml:"flush-interval"`
-			ProxyURL          string                 `yaml:"proxy-url"`
-			TLSInsecure       bool                   `yaml:"tls-insecure"`
-			TLSMinVersion     string                 `yaml:"tls-min-version"`
-			CAFile            string                 `yaml:"ca-file"`
-			CertFile          string                 `yaml:"cert-file"`
-			KeyFile           string                 `yaml:"key-file"`
-			ChannelBufferSize int                    `yaml:"chan-buffer-size"`
-		} `yaml:"scalyrclient"`
-		RedisPub struct {
-			Enable            bool   `yaml:"enable"`
-			RemoteAddress     string `yaml:"remote-address"`
-			RemotePort        int    `yaml:"remote-port"`
-			SockPath          string `yaml:"sock-path"` // deprecated
-			RetryInterval     int    `yaml:"retry-interval"`
-			Transport         string `yaml:"transport"`
-			TLSSupport        bool   `yaml:"tls-support"` // deprecated
-			TLSInsecure       bool   `yaml:"tls-insecure"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CAFile            string `yaml:"ca-file"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			Mode              string `yaml:"mode"`
-			TextFormat        string `yaml:"text-format"`
-			PayloadDelimiter  string `yaml:"delimiter"`
-			BufferSize        int    `yaml:"buffer-size"`
-			FlushInterval     int    `yaml:"flush-interval"`
-			ConnectTimeout    int    `yaml:"connect-timeout"`
-			RedisChannel      string `yaml:"redis-channel"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"redispub"`
-		KafkaProducer struct {
-			Enable            bool   `yaml:"enable"`
-			RemoteAddress     string `yaml:"remote-address"`
-			RemotePort        int    `yaml:"remote-port"`
-			RetryInterval     int    `yaml:"retry-interval"`
-			TLSSupport        bool   `yaml:"tls-support"`
-			TLSInsecure       bool   `yaml:"tls-insecure"`
-			TLSMinVersion     string `yaml:"tls-min-version"`
-			CAFile            string `yaml:"ca-file"`
-			CertFile          string `yaml:"cert-file"`
-			KeyFile           string `yaml:"key-file"`
-			SaslSupport       bool   `yaml:"sasl-support"`
-			SaslUsername      string `yaml:"sasl-username"`
-			SaslPassword      string `yaml:"sasl-password"`
-			SaslMechanism     string `yaml:"sasl-mechanism"`
-			Mode              string `yaml:"mode"`
-			BufferSize        int    `yaml:"buffer-size"`
-			FlushInterval     int    `yaml:"flush-interval"`
-			ConnectTimeout    int    `yaml:"connect-timeout"`
-			Topic             string `yaml:"topic"`
-			Partition         int    `yaml:"partition"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-			Compression       string `yaml:"compression"`
-		} `yaml:"kafkaproducer"`
-		FalcoClient struct {
-			Enable            bool   `yaml:"enable"`
-			URL               string `yaml:"url"`
-			ChannelBufferSize int    `yaml:"chan-buffer-size"`
-		} `yaml:"falco"`
-	} `yaml:"loggers"`
-
+	Global               ConfigGlobal       `yaml:"global"`
+	Collectors           ConfigCollectors   `yaml:"collectors"`
+	IngoingTransformers  ConfigTransformers `yaml:"collectors-transformers"`
+	Loggers              ConfigLoggers      `yaml:"loggers"`
 	OutgoingTransformers ConfigTransformers `yaml:"loggers-transformers"`
-
-	Multiplexer struct {
-		Collectors []MultiplexInOut  `yaml:"collectors"`
-		Loggers    []MultiplexInOut  `yaml:"loggers"`
-		Routes     []MultiplexRoutes `yaml:"routes"`
-	} `yaml:"multiplexer"`
+	Multiplexer          ConfigMultiplexer  `yaml:"multiplexer"`
 }
 
 func (c *Config) SetDefault() {
@@ -835,7 +343,6 @@ func (c *Config) SetDefault() {
 
 	// Transformers for loggers
 	c.OutgoingTransformers.SetDefault()
-
 }
 
 func (c *Config) GetServerIdentity() string {
@@ -893,12 +400,6 @@ func LoadConfig(configPath string) (*Config, error) {
 
 func GetFakeConfig() *Config {
 	config := &Config{}
-	config.SetDefault()
-	return config
-}
-
-func GetFakeConfigTransformers() *ConfigTransformers {
-	config := &ConfigTransformers{}
 	config.SetDefault()
 	return config
 }
