+++ pkgconfig/multiplexer.go
@@ -1,0 +1,18 @@
+package pkgconfig
+
+type ConfigMultiplexer struct {
+	Collectors []MultiplexInOut  `yaml:"collectors"`
+	Loggers    []MultiplexInOut  `yaml:"loggers"`
+	Routes     []MultiplexRoutes `yaml:"routes"`
+}
+
+type MultiplexInOut struct {
+	Name       string                 `yaml:"name"`
+	Transforms map[string]interface{} `yaml:"transforms"`
+	Params     map[string]interface{} `yaml:",inline"`
+}
+
+type MultiplexRoutes struct {
+	Src []string `yaml:"from,flow"`
+	Dst []string `yaml:"to,flow"`
+}
